test: split FloatingRail unit specs
